kommunikations test og lidt omrokering
